requirements-docker.txt
# Runtime dependencies for Docker containers (Python 3.13)
# Only direct dependencies are pinned. Transitive deps are resolved by pip.
# Dev/test/lint tools are excluded — see requirements.txt for the full set.

# --- Core framework ---
crewai[litellm]==1.10.1
chainlit==2.9.6
fastapi>=0.129.0
uvicorn>=0.40.0

# --- LLM providers ---
openai>=1.83.0
anthropic>=0.73.0
litellm>=1.74.9

# --- Async / networking ---
aiohttp>=3.9.0
aiosmtplib>=2.0.0
httpx>=0.28.0
websockets>=13.0

# --- Data / persistence ---
asyncpg>=0.30.0
SQLAlchemy>=2.0.40
alembic>=1.18.0
redis>=7.0.0
celery>=5.4.0
pandas>=2.2.0
numpy>=1.26.4
openpyxl>=3.1.5

# --- Auth / crypto ---
Authlib>=1.6.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
cryptography>=42.0.0

# --- Observability ---
opentelemetry-api>=1.34.0
opentelemetry-sdk>=1.34.0
opentelemetry-exporter-otlp-proto-http>=1.34.0
opentelemetry-exporter-otlp-proto-grpc>=1.34.0
opentelemetry-instrumentation-redis>=0.50b0
opentelemetry-instrumentation-requests>=0.50b0
opentelemetry-instrumentation-sqlalchemy>=0.50b0
prometheus_client>=0.20.0
structlog>=25.0.0
traceloop-sdk>=0.30.0

# --- Document processing ---
pdfminer.six>=20221105
pdfplumber>=0.11.0
pillow>=10.0.0
opencv-python-headless>=4.10.0

# --- ML ---
scikit-learn>=1.5.0
nltk>=3.9.0
onnxruntime>=1.18.0

# --- Scheduling / infra ---
APScheduler>=3.10.0

# --- Serialization / config ---
pydantic>=2.10.0
pydantic-settings>=2.10.0
PyYAML>=6.0.0
orjson>=3.10.0
ruamel.yaml>=0.18.0
python-dotenv>=1.0.0

# --- Utilities ---
backoff>=2.2.0
Jinja2>=3.1.0
requests>=2.31.0
rich>=13.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# --- Testing (used by agents at runtime) ---
pytest>=9.0.0
Faker>=40.0.0
playwright>=1.40.0

# --- Python 3.13 compat ---
audioop-lts>=0.2.0
